Fix compiler error when using clang:
src/misc/AtomicString.cpp:175:28: error: non-constant-expression
 cannot be narrowed from type 'int' to 'unsigned int' in initializer
 list [-Wc++11-narrowing]
    UCharBuffer buf = { s, length };
                           ^~~~~~
src/misc/AtomicString.cpp:175:28: note: override this message by
 inserting an explicit cast
    UCharBuffer buf = { s, length };
                           ^~~~~~
                           static_cast<unsigned int>( )

REVIEW: 116541

+++ src/misc/AtomicString.cpp
@@ -172,7 +172,7 @@ DOMStringImpl *AtomicString::add(const QChar *s, int length)
     }
 
     init();
-    UCharBuffer buf = { s, length };
+    UCharBuffer buf = { s, static_cast<uint>(length) };
     std::pair<HashSet<DOMStringImpl *>::iterator, bool> addResult = stringTable->add<UCharBuffer, UCharBufferTranslator>(buf);
     if (!addResult.second) {
         return *addResult.first;
@@ -196,7 +196,7 @@ DOMStringImpl *AtomicString::add(const QChar *s)
     }
 
     init();
-    UCharBuffer buf = {s, length};
+    UCharBuffer buf = {s, static_cast<uint>(length)};
     std::pair<HashSet<DOMStringImpl *>::iterator, bool> addResult = stringTable->add<UCharBuffer, UCharBufferTranslator>(buf);
     if (!addResult.second) {
         return *addResult.first;
